+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/userinfo/UserInfoBridgeDispatcher.kt
@@ -34,7 +34,7 @@ abstract class UserInfoBridgeDispatcher {
      * You can send user name or throw an [Exception] from this method by passing a [Result].
      */
     open fun getUserName(
-        callback: (result: Result<String?>) -> Unit
+        callback: (userName: Result<String?>) -> Unit
     ) {
         throw MiniAppSdkException("The `UserInfoBridgeDispatcher.getUserName` $NO_IMPL")
     }
@@ -55,12 +55,16 @@ abstract class UserInfoBridgeDispatcher {
      * You can send profile photo url or throw an [Exception] from this method by passing a [Result].
      */
     open fun getProfilePhoto(
-        callback: (result: Result<String?>) -> Unit
+        callback: (profilePhoto: Result<String?>) -> Unit
     ) {
         throw MiniAppSdkException("The `UserInfoBridgeDispatcher.getProfilePhoto` $NO_IMPL")
     }
 
     /** Get access token from host app. **/
+    @Deprecated(
+        "This function has been deprecated.",
+        ReplaceWith("getAccessToken(miniAppId: String, callback: (result: Result<TokenData?>) -> Unit)")
+    )
     open fun getAccessToken(
         miniAppId: String,
         onSuccess: (tokenData: TokenData) -> Unit,
@@ -69,13 +73,37 @@ abstract class UserInfoBridgeDispatcher {
         throw MiniAppSdkException("The `UserInfoBridgeDispatcher.getAccessToken` $NO_IMPL")
     }
 
+    /** Get access token from host app.
+     * You can send access token or throw an [Exception] from this method by passing a [Result].
+     * */
+    open fun getAccessToken(
+        miniAppId: String,
+        callback: (accessToken: Result<TokenData?>) -> Unit
+    ) {
+        throw MiniAppSdkException("The `UserInfoBridgeDispatcher.getAccessToken` $NO_IMPL")
+    }
+
     /**
      * Get contacts from host app.
      * You can also throw an [Exception] from this method to pass an error message to the mini app.
      */
+    @Deprecated(
+        "This function has been deprecated.",
+        ReplaceWith("getContacts(callback: (result: Result<ArrayList<Contact>?>) -> Unit)")
+    )
     open fun getContacts(
         onSuccess: (contacts: ArrayList<Contact>) -> Unit,
         onError: (message: String) -> Unit
+    ) {
+        throw MiniAppSdkException("The `UserInfoBridgeDispatcher.getContacts` $NO_IMPL")
+    }
+
+    /**
+     * Get contacts from host app.
+     * You can send contacts or throw an [Exception] from this method by passing a [Result].
+     */
+    open fun getContacts(
+        callback: (contacts: Result<ArrayList<Contact>?>) -> Unit
     ) {
         throw MiniAppSdkException("The `UserInfoBridgeDispatcher.getContacts` $NO_IMPL")
     }
@@ -90,6 +118,7 @@ abstract class UserInfoBridgeDispatcher {
         this.miniAppId = miniAppId
     }
 
+    /** region: user name */
     internal fun onGetUserName(callbackId: String) {
         if (customPermissionCache.hasPermission(
                 miniAppId,
@@ -120,7 +149,8 @@ abstract class UserInfoBridgeDispatcher {
             )
     }
 
-    private fun getUserNameSync(callbackId: String) {
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    internal fun getUserNameSync(callbackId: String) {
         try {
             val name = getUserName()
             if (name.isNotEmpty()) bridgeExecutor.postValue(callbackId, name)
@@ -132,7 +162,9 @@ abstract class UserInfoBridgeDispatcher {
             bridgeExecutor.postError(callbackId, "$ERR_GET_USER_NAME ${e.message}")
         }
     }
+    /** end region */
 
+    /** region: profile photo */
     internal fun onGetProfilePhoto(callbackId: String) {
         if (customPermissionCache.hasPermission(
                 miniAppId,
@@ -163,7 +195,8 @@ abstract class UserInfoBridgeDispatcher {
             )
     }
 
-    private fun getProfilePhotoSync(callbackId: String) {
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    internal fun getProfilePhotoSync(callbackId: String) {
         try {
             val photoUrl = getProfilePhoto()
             if (photoUrl.isNotEmpty()) bridgeExecutor.postValue(callbackId, photoUrl)
@@ -175,33 +208,68 @@ abstract class UserInfoBridgeDispatcher {
             bridgeExecutor.postError(callbackId, "$ERR_GET_PROFILE_PHOTO ${e.message}")
         }
     }
+    /** end region */
 
-    internal fun onGetAccessToken(callbackId: String) = try {
-        val successCallback = { accessToken: TokenData ->
-            bridgeExecutor.postValue(callbackId, Gson().toJson(accessToken))
+    /** region: access token */
+    internal fun onGetAccessToken(callbackId: String) {
+        try {
+            // retrieve access token using Result
+            val resultCallback = { result: Result<TokenData?> ->
+                if (result.isSuccess) bridgeExecutor.postValue(
+                    callbackId,
+                    Gson().toJson(result.getOrNull())
+                )
+                else if (result.isFailure) bridgeExecutor.postError(
+                    callbackId,
+                    "$ERR_GET_ACCESS_TOKEN ${result.exceptionOrNull()?.message}"
+                )
+            }
+            getAccessToken(miniAppId, resultCallback)
+        } catch (e: Exception) {
+            if (e.message.toString().contains(NO_IMPL)) getAccessTokenWithoutResult(callbackId)
+            else bridgeExecutor.postError(callbackId, "$ERR_GET_ACCESS_TOKEN ${e.message}")
         }
-        val errorCallback = { message: String ->
-            bridgeExecutor.postError(callbackId, "$ERR_GET_ACCESS_TOKEN $message")
-        }
-
-        getAccessToken(miniAppId, successCallback, errorCallback)
-    } catch (e: Exception) {
-        bridgeExecutor.postError(callbackId, "$ERR_GET_ACCESS_TOKEN ${e.message}")
     }
 
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    internal fun getAccessTokenWithoutResult(callbackId: String) {
+        try {
+            val successCallback = { accessToken: TokenData ->
+                bridgeExecutor.postValue(callbackId, Gson().toJson(accessToken))
+            }
+            val errorCallback = { message: String ->
+                bridgeExecutor.postError(callbackId, "$ERR_GET_ACCESS_TOKEN $message")
+            }
+            getAccessToken(miniAppId, successCallback, errorCallback)
+        } catch (e: Exception) {
+            bridgeExecutor.postError(callbackId, "$ERR_GET_ACCESS_TOKEN ${e.message}")
+        }
+    }
+    /** end region */
+
+    /** region: contacts */
     internal fun onGetContacts(callbackId: String) = try {
         if (customPermissionCache.hasPermission(
                 miniAppId, MiniAppCustomPermissionType.CONTACT_LIST
             )
         ) {
-            val successCallback = { contacts: ArrayList<Contact> ->
-                bridgeExecutor.postValue(callbackId, Gson().toJson(contacts))
+            try {
+                // retrieve contacts using Result
+                val resultCallback = { result: Result<ArrayList<Contact>?> ->
+                    if (result.isSuccess) bridgeExecutor.postValue(
+                        callbackId,
+                        Gson().toJson(result.getOrNull())
+                    )
+                    else if (result.isFailure) bridgeExecutor.postError(
+                        callbackId,
+                        "$ERR_GET_CONTACTS ${result.exceptionOrNull()?.message}"
+                    )
+                }
+                getContacts(resultCallback)
+            } catch (e: Exception) {
+                if (e.message.toString().contains(NO_IMPL)) getContactsWithoutResult(callbackId)
+                else bridgeExecutor.postError(callbackId, "$ERR_GET_CONTACTS ${e.message}")
             }
-            val errorCallback = { message: String ->
-                bridgeExecutor.postError(callbackId, "$ERR_GET_CONTACTS $message")
-            }
-
-            getContacts(successCallback, errorCallback)
         } else {
             bridgeExecutor.postError(
                 callbackId,
@@ -211,6 +279,22 @@ abstract class UserInfoBridgeDispatcher {
     } catch (e: Exception) {
         bridgeExecutor.postError(callbackId, "$ERR_GET_CONTACTS ${e.message}")
     }
+
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    internal fun getContactsWithoutResult(callbackId: String) {
+        try {
+            val successCallback = { contacts: ArrayList<Contact> ->
+                bridgeExecutor.postValue(callbackId, Gson().toJson(contacts))
+            }
+            val errorCallback = { message: String ->
+                bridgeExecutor.postError(callbackId, "$ERR_GET_CONTACTS $message")
+            }
+            getContacts(successCallback, errorCallback)
+        } catch (e: Exception) {
+            bridgeExecutor.postError(callbackId, "$ERR_GET_CONTACTS ${e.message}")
+        }
+    }
+    /** end region */
 
     @VisibleForTesting
     internal companion object {